envoy: Reduce error logging
[ upstream commit 430e535b44aa72a95032e91b4bcd4281e6321448 ]

gRPC stream on the server side gets cancelled if the client side
cancels the stream. This happens when the Envoy listener is being shut
down, which happens when the policy is changed to not contain L7 rules
any more. As this is part of the normal operation, change these logs
to the debug level.

This should remove these from the CI output:

Top 5 errors/warnings:
error while receiving request from xDS stream
xDS stream context canceled

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>
Signed-off-by: Ian Vernon <ian@cilium.io>