eni: Retry on attachment index conflict
Allocation is delayed until metadata is synced but it can still be oudated when
an ENI is attached to an instance. Detect the error returned which indicates
that the index is already in use. Retry with the next available index up to
a maximum number of retries.

ENI creation is expensive, deletion due to index conflict should be avoided.

Signed-off-by: Thomas Graf <thomas@cilium.io>